Refactor ui code structure
