+++ src/ui/components/StepList.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import styled from 'styled-components';
 import { inject, observer } from 'mobx-react';
 
-import { handleAnchor, isClientOrServer, rem } from '../utils/common';
-import Store from './store';
+import { handleAnchor, isClientOrServer, rem } from '../utils';
+import Store from '../store';
 
 export interface StepListProps {
   commits: Commit[];
