+++ src/ui/store.ts
@@ -2,7 +2,7 @@ import { observable, action, computed } from 'mobx';
 import classnames from 'classnames';
 import fetch from 'isomorphic-fetch';
 
-import { handleAnchor } from '../utils/common';
+import { handleAnchor } from './utils';
 
 class Store {
   @observable
